+++ src/ValidationPackage.Model/index.js
@@ -1,0 +1,14 @@
+export * from "./Author.js";
+export * from "./Cwl.js";
+export const CwlPrimitive = Object.freeze({
+  Boolean: 0,
+  Int: 1,
+  Long: 2,
+  Float: 3,
+  Double: 4,
+  String: 5
+});
+export * from "./OntologyAnnotation.js";
+export * from "./SemanticVersion.js";
+export * from "./ValidationPackageIdentity.js";
+export * from "./ValidationPackageMetadata.js";